Group batting mod pages
